La til godkjenning av kontrakter

+++ ShareBNB/profile_page/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponseForbidden
 
-from apartments.models import Apartment
+from apartments.models import Apartment, Contract
 from authentication.models import Profile
 
 
@@ -15,6 +15,25 @@ def profile_view(request):
         else:
             return redirect('landing-page')
     else: # POST
-        user = Profile.objects.get(pk=request.user.pk)
-        user.delete()
-        return redirect('landing-page')
+        contract_id = request.POST.get("contract_id")
+
+        if not contract_id==None:
+            #Godkjenner kontrakten
+            contract = Contract.objects.get(pk=contract_id)
+            contract.pending = False;
+            contract.save()
+
+            #Viser siden dersom brukeren fremdeles er logget inn
+            if request.user.is_authenticated:
+                context = {
+                    'my_apartments': Apartment.objects.filter(owner=request.user)
+                }
+                return render(request, 'profile_page/profile-page.html', context)
+            else:
+                return redirect('landing-page')
+
+
+        else:
+            user = Profile.objects.get(pk=request.user.pk)
+            user.delete()
+            return redirect('landing-page')
